requirements.txt
mutagen>=1.45.0
PyYAML>=6.0
pydantic>=2.0.0
pyacoustid>=1.3.0
shazamio>=0.8.0
yt-dlp>=2024.0.0
Pillow>=10.0.0
requests>=2.31.0

# Optional dependencies for DJ mix segmentation
# Install with: pip install -r requirements.txt librosa numpy
# librosa>=0.10.0
# numpy>=1.24.0
